+++ resources/js/pages/demos/split-button/avatar.tsx
@@ -1,0 +1,42 @@
+import {
+  CreditCardIcon,
+  Logout02Icon,
+  NotificationSquareIcon,
+  UserCircleIcon
+} from '@hugeicons/core-free-icons'
+import { Pressable } from 'react-aria-components'
+import { DemoContainer } from '@/components/docs/DemoContainer'
+
+import { Avatar } from '@/components/twc-ui/avatar'
+import { DropdownButton } from '@/components/twc-ui/dropdown-button'
+import { MenuHeader, MenuItem } from '@/components/twc-ui/menu'
+export const Demo = () => {
+  return (
+    <DemoContainer className="w-max-lg items-start gap-4 py-12">
+      <DropdownButton
+        menuClassName="min-w-64"
+        triggerElement={
+          <Pressable>
+            <Avatar src="https://github.com/shadcn.png" initials="DS" size="lg" role="button" />
+          </Pressable>
+        }
+      >
+        <MenuHeader>
+          <div className="flex items-center gap-2">
+            <Avatar src="https://github.com/shadcn.png" initials="DS" size="lg" />
+            <div className="text-base">
+              Danny
+              <div className="font-normal text-sm">danny@example.com</div>
+            </div>
+          </div>
+        </MenuHeader>
+        <MenuItem icon={UserCircleIcon} title="Account" ellipsis />
+        <MenuItem icon={CreditCardIcon} title="Billing" ellipsis />
+        <MenuItem icon={NotificationSquareIcon} title="Notifications" ellipsis separator />
+        <MenuItem icon={Logout02Icon} title="Logout" />
+      </DropdownButton>
+    </DemoContainer>
+  )
+}
+
+export default Demo